requirements.txt
Flask==2.0.2
peewee==3.14.8
python-dotenv==0.19.2
pymysql==1.0.2
requests~=2.26.0